+++ tests/test_fuzz.py
@@ -8,8 +8,10 @@ from array import array
 from rapidfuzz import fuzz_py, fuzz_cpp, utils
 from rapidfuzz.distance import ScoreAlignment
 
+
 def isclose(a, b, rel_tol=1e-09, abs_tol=0.0):
     return abs(a - b) <= max(rel_tol * max(abs(a), abs(b)), abs_tol)
+
 
 class fuzz:
     @staticmethod